requirements.txt
cohere==1.3.1
datasets==2.1.0
hydra-core==1.2.0
numpy==1.22.3
omegaconf==2.1.1
openai==0.15.0
scipy==1.7.1
tqdm==4.62.3
transformers==4.18.0
wandb==0.12.11
